Ajustei o banco de dados e a tela de Venda

+++ database/migrations/2021_03_16_005231_criar_tabela_vendas.php
@@ -18,7 +18,7 @@ class CriarTabelaVendas extends Migration
             $table->integer('fkCliente');
             $table->integer('fkProduto');
             $table->integer('qtd');
-            $table->float('valor');
+            $table->double('valor', 8, 2);
             $table->timestamps();
         });
     }
